Modes triggers with full testing
